WIP: split keymap,small config modules and tree
